chore(oidc-client): move callback handler to another file
