+++ packages/oidc-client/src/server/middleware/callback.ts
@@ -1,0 +1,66 @@
+import { NextResponse } from "next/server";
+import { MiddlewareHandlerOptions } from "../middleware";
+import { internalExchangeTokenForSession } from "@lib/core";
+import { Config } from "@/lib/config";
+import { Session } from "@/lib/types";
+
+export class CallbackError extends Error {
+  constructor(
+    readonly name: string,
+    readonly message: string,
+  ) {
+    super(message);
+  }
+}
+
+export const paramsError = () =>
+  new CallbackError("invalid-params", "code and redirectURI should be filled");
+export const exchangeError = (reason: string) =>
+  new CallbackError("failed-exchange", reason);
+
+export const callbackHandler = async (
+  params: URLSearchParams,
+  config: Config,
+  options?: MiddlewareHandlerOptions,
+) => {
+  const code = params.get("code");
+  const redirectURI = params.get("redirect_uri");
+  if (!code || !redirectURI) {
+    throw paramsError();
+  }
+
+  const session = await internalExchangeTokenForSession(config, code);
+  if ("error" in session) {
+    throw exchangeError(session.error);
+  }
+  options?.prepend &&
+    (await options.prepend({
+      token: session.access_token,
+      userID: session.user_id,
+    }));
+
+  const redirection = NextResponse.redirect(config.appUrl(redirectURI));
+  redirection.cookies.set(
+    buildCookieEntry(session, "tailor.token", "access_token"),
+  );
+  return redirection;
+};
+
+const buildCookieEntry = <const T extends keyof Session>(
+  session: Session,
+  name: string,
+  value: T,
+) => {
+  // Use `expires_in` field from Tailor Platform auth service which comes as seconds
+  const expires = session.expires_in * 1000;
+
+  // Use the strictest cookie here
+  return {
+    name,
+    value: session[value],
+    sameSite: "strict" as const,
+    httpOnly: true,
+    secure: true,
+    expires,
+  };
+};
